requirements.txt
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cssselect2==0.7.0
distro==1.9.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
jiter==0.5.0
markdown-it-py==3.0.0
markdown2==2.5.0
markdown_pdf==1.3
mdurl==0.1.2
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyphen==0.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tinycss2==1.3.0
tokenizers==0.20.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zopfli==0.2.3
